Fix viewport mesh rendering and add wireframe overlay
- Fixed vertex layout mismatch (32 bytes: pos+normal+uv)
- Added wireframe pipeline with subtle 1-pixel lines for 3D depth perception
- Added FillMode enum and LineWidth to RasterizerState
- Metal pipeline now supports triangle fill mode
- Removed debug logging spam
- Removed unused Horizon Lighting buffers and code
- Fixed shader loading pattern matching for fs_wireframe

+++ BlueSkyEngine/Editor/Program.cs
@@ -381,7 +381,7 @@ class Program
                 var camPos = _viewport.GetCameraPositionNumerics();
 
                 _viewportRenderer!.Render(cmd, view, proj, camPos,
-                    vpXPx, vpYPx, vpWPx, vpHPx, _deltaTime);
+                    (int)vpXPx, (int)vpYPx, (int)vpWPx, (int)vpHPx, _deltaTime);
 
                 // Reset viewport/scissor back to full window for UI
                 cmd.SetViewport(new NotBSRenderer.Viewport { X = 0, Y = 0, Width = w, Height = h, MinDepth = 0, MaxDepth = 1 });
